Refactor: applying refactoring changes to dependent modules and tests
