+++ nlp_profiler/granular_features/chars_and_spaces.py
