+++ nlp_profiler/granular_features/emojis.py
