+++ nlp_profiler/granular_features/stop_words.py
