+++ tests/granular/test_dates.py
@@ -2,7 +2,8 @@ import numpy as np
 import pytest
 
 from nlp_profiler.constants import NaN
-from nlp_profiler.dates import gather_dates, count_dates  # noqa
+from nlp_profiler.granular_features.dates \
+    import gather_dates, count_dates  # noqa
 
 text_with_date = "Todays date is 04/28/2020 for format mm/dd/yyyy, not 28/04/2020"
 text_with_dates = "Todays date is 28/04/2020 and tomorrow's date is 29/04/2020"
