+++ tests/granular/test_non_alphanumeric.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from nlp_profiler.constants import NaN
-from nlp_profiler.non_alphanumeric \
+from nlp_profiler.granular_features.non_alphanumeric \
     import gather_non_alpha_numeric, count_non_alpha_numeric  # noqa
 
 text_with_a_number = '2833047 people live in this area'
